Addition of TCP/Single-core implementation

+++ src/hydrophones/AnalogMeasurement.h
@@ -1,13 +1,15 @@
 #ifndef MEASUREMENT_H
 #define MEASUREMENT_H
 
+#include <cstdint>
+
 struct AnalogMeasurement
 {
     /*
-     * Specifies the time at which the sample was taken (in microseconds)
-     * (relative).
+     * Specifies the time at which the sample was taken in seconds. This is a
+     * relative time from the start of sampling.
      */
-    uint64_t timestamp;
+    float timestamp;
 
     /*
      * The actual analog sample measurement.
